dashboard/requirements.txt
streamlit
psycopg2-binary
google-cloud-logging
google-auth